+++ styles/globals.scss
@@ -6,7 +6,7 @@
   --background_3: #2f3136;
   --background_4: #40444b;
   --background_5: #666d76;
-  --disabled_filter: grayscale(100%) brightness(50%);
+  --disabled_filter: #373737bc;
 }
 
 @font-face {
